Add app functionality
Update with actual functionality of app, including whatsForDinner.py. This is a standalone python module that takes user interaction and recommends dinner options.

Add other directories and init file which are really for the web app, but already in the directory.

If you want to use a standalone python script, use this version of whatsForDinner.py